Subscribe customer to default payment plan after charging on stripe.
A new customer is created using the Stripe API and the customer
is subscribed to the default payment plan defined in settings.

In addition, SHOP_STRIPE_API_VERSION is used to specify which
Stripe API version to use. Currently, djangoshop-stripe requires
a version higher than '2015-02-16' due to backward-incompatible
changes made on the charge object (charge.status returns 'succeeded'
instead of 'paid').